feat(login): écran unique avec 2 composants de connexion
Ils sont identiques aux précédents, mais sur une même page.
