+++ src/pages/exploitation/index.vue
@@ -97,7 +97,7 @@ meta:
           </router-link>
         </p>
         <p v-else>
-          <router-link class="fr-btn fr-btn--secondary" to="/exploitation/login">
+          <router-link class="fr-btn fr-btn--secondary" to="/login">
             Me connecter à CartoBio
           </router-link>
         </p>
